@@ -4,12 +4,12 @@ version: 0.1.0
 homepage: https://github.com/edunuzzi/flutter_jigsaw_puzzle
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
-  image: ^2.1.12
-  carousel_slider: ^3.0.0
+  image: ^3.0.5
+  carousel_slider: ^4.0.0
   clip_shadow: ^0.2.1
